Use pkg-config for libdrm in eglfs_kms_egldevice backend if possible
Without pkg-config, /usr/include/libdrm is not added to the include
search path so xf86drm.h may be unable to include drm.h.

This change is being applied directly to 5.6 because it's not
applicable to 5.9 or dev. Those versions of Qt have moved to using
QMAKE_USE which will already use pkg-config if available.

Change-Id: I875c1e5f1b3cce63fd6b529d9d8ac97fd5bb56e8
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>