QHash: make MSVC happy about the iterators passed to is_permutation
MSVC warns about iterators being passed to certain Standard Library
algorithms. dbd55cdaf367bdc9d6774bcb9927cbe19f18065f introduced
a usa of std::is_permutation in a public header, which is causing
such a warning to be emitted.

To suppress the warning, Microsoft suggests to either use the 4-arg
std::is_permutation overload (which however is not available
in MSVC 2013) or to use a Standard Library extension, which we are
already using elsewhere in Qt to deal with the same problem. However,
that extension requires the iterator to be moved by size_t quantities,
which isn't the case for QHash::iterator, and therefore generates
more warnings about loss of precision (size_t -> int).

Therefore, go with the 4-arg std::is_permutation, only on MSVC >= 2015.

Change-Id: Idfcff28d14e0f1fde5d77f1deb9eec27c87ff5cd
Task-number: QTBUG-61902
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>